+++ CHANGELOG.md
@@ -10,6 +10,8 @@
 ### Fixed
 
 * Calculation of text related [HexaPDF::Layout::Style] values for Type3 fonts
+* [HexaPDF::Encryption::SecurityHandler#encrypt_string] to either return a
+  dupped or encrypted string
 
 
 ## 0.24.2 - 2022-08-31